ТЗ-15 add item requests

+++ src/main/java/ru/practicum/shareit/booking/dto/BookingRequestDto.java
@@ -25,5 +25,5 @@ public class BookingRequestDto {
     }
 
     @NotNull(message = "ID вещи не может быть пустым")
-    private long itemId;
+    private int itemId;
 }
